Mesh shader addition + Transfer specific buffers for vertex data passing to shader + added bufferMemory barriers for smooth transferring the data + optimized mesh with less meshlet counts to mesh shader

+++ src/shaders/mesh.vert
@@ -5,9 +5,9 @@
 #extension GL_EXT_shader_explicit_arithmetic_types: require
 
 struct Vertex{
-    float vx, vy, vz;
+    float16_t vx, vy, vz;
     uint8_t nx, ny, nz, nw;
-    float tu, tv;
+    float16_t tu, tv;
 };
 
 layout(binding=0) readonly buffer Vertices{
